extract vcvars test
